don start the zabbix agent as root
This was very risky. We started the agent always as root. If there is
another username provided in the config, the agent will drop priviledges
and change to the provided user. Also we use the same user for
zabbix-server + zabbix-agent. This commit changes a few things. The
use for the agent is now configureable, on archlinux we use a dedicated
user for the agent (provided by the upstream package). Also we don't
start the agent as root anywhere.