handle changing defaults on diff. zabbix versions
this is a really really nasty bug. A few of the default values we
provide depend on the zabbix version. Their default value is different
based on the zabbix version. We deal with this in the params.pp, but
that codeblock is always ignored if you provide a zabbix version by
yourself. This is now fixed and proven to work by acceptance tests.