Enable doc installations in centos containers
The official CentOS docker container disables the installation of
RPM docs in the image via the yum.conf file. The added code will check
beaker host platform for EL 7 and host hypervisor for docker and
re-enable the installation of docs files.

This change is due to zabbix-server-pgsql tests failing as it relies
on a sample file provided in the RPM docs directory.