Update module dependencies
Fixes #453

We test against the latest versions and we've had reports that people
are successfully using this module with recent versions of
apache/concat/apt etc.

Some dependencies should probably be removed as they are 'soft'
dependencies.  This would at least need a README update though.  For
now, this commit is probably good enough to be worth merging.