Do not guess ListenIP Fixes #473
This commit changes the behaviour if the $listen_ip argument is not
given. In that case ListenIP is omitted from the agent configuration.
Having no ListenIP in the configuration ensures the agent listens on
both IPv4 and IPv6.
Also added a spec test for the case were ListenIP is left empty.

Signed-off-by: Olivier van der Toorn <oliviervdtoorn@gmail.com>