Zabbix.com provides a repository for Raspbian Linux under http://repo.zabbix.com/zabbix/<version>/raspbian/
The puppet-zabbix module uses the fact os.name for determining the URL of the used repository which is "Debian" for Raspbian Linux. The Debian repositiry is not providing "armhf" deb packages so the installation of zabbix fails.

This patch is an if condition which sets the variable $operatingsystem to 'raspbian' if the fact os.distro.id is 'Raspbian'. Otherwise the behaviour is not modified.