Reduce configuration on agent service
This avoids setting parameters as much as possible.

hasstatus and hasrestart have defaulted to true since Puppet 2.7 so
there's no need to set those.

The systemd fact is defined as $fact['service_provider'] == 'systemd' so
there's no need to set that either.

Then there's AIX left. It sets it this was because it allows users to
override it in site.pp via service defaults. It now only sets the
service_provider and not the path. This path is set in the provider
since Puppet 4.6.0[1].

https://github.com/puppetlabs/puppet/commit/34e45ab3f562d1029bfeaafe9a11757ea3b3e92c