Drop support for Puppet 6
This module install gems using the puppet_gem provider in the gempath of
the version of Ruby bundled with Puppet.  Puppet 6 bundle Ruby 2.5 which
has reached EoL on 2022-04-05, and the gems the module needs depend on
gems that do not offer support for this version of Ruby anymore.

This break CI and the situation has no reason to improve in the future,
so stop advertising support for Puppet 6.