Use include postgresql::server in zabbix::database
This is safe to do because it can be called even after an explicit class
statement. The following code is legal.

    class { 'postgresql::server':
    }
    include postgresql::server

It could make a difference for a remote database, but due to the usage
of postgresql::server::pg_hba_rule that was already invalid.

The explicit require is also not needed, since postgresql::server::db
already does this where needed.

The end result is that the module is easier to use.