Set header for authorization
On Zabbix version 6.4 and newer there has been a change in the way
authentication. Using a bearer token would lead to issues with
autheticating using the API as the header did not seem the get passed to
the application.

This seems to have been known since the release of 6.4 and a bug report
is still open. But the bug report also already contained a
workaround/fix (https://support.zabbix.com/browse/ZBX-22952).